pyproject.toml
[project]
name = "thinkdepthai_deep_research"
version = "0.1.3"
description = "ThinkDepth.ai Deep Research"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.0.5",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "langchain-azure-ai>=1.0.0",
    "langchain-anthropic>=1.3.0",
    "langchain-core>=1.2.5",
    "langchain_community>=0.3.27",
    "langchain_tavily>=0.2.7",
    "langchain_mcp_adapters>=0.1.9",
    "typing_extensions>=4.15.0",
    "pydantic>=2.12.5",
    "rich>=14.2.0",
    "PyYAML>=6.0.3",
    "tavily-python>=0.7.17",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.1",
    "ruff>=0.6.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
    "build>=1.2.2",
]

# Optional: keep notebooks out of the runtime dependency set.
# If you want them installable on demand:
notebook = [
    "jupyter>=1.1.1",
    "ipykernel>=7.1.0",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = false

[tool.setuptools.packages.find]
where = ["src"]
include = ["deep_research*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
deep_research = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"
